Revamp PubSubJedisClient
This commit drastically improves PubSubJedisClient. First and foremost,
this commit adds the ability of PubSubJedisClient to reconnect to its
target redis instance if the connection fails or is interrupted at any
point. It will continue trying to reconnect periodically until it is
destroyed. When connection is regained, any channels that were
previously subscribed to or where an attempt was made to subscribe while
the connection was down are subscribed to again. 

Additionally, This commit improves the exception handling, efficiency,
logic, logging, and general safety of PubSubJedisClient.