Add publishing feedback and messager grouping.
This is a major commit that adds a number of new features and makes a
non-backwards-compatible change the PubSubMessager interface.

Summary of changes:
- Publish interface now returns a ListenableFuture object which gives
limited feedback on whether the publish worked or not.
- Combined PubSubLibrarySubscription and Publisher interfaces into a
single PubSubLibraryClient interface. It is just simpler in practice. If
they need to be separated, it can be done in implementation.
- Added the ability to group pub/sub messaging connections/backends
together with the RedundantClient class, for the sake of load balancing
and fail tolerance. If both sender and receiver subscribe to all
channels that might be published to, then it does not matter how the
messages get where they are going. Added a few different modes for load
balancing. 
- Added guava and ExpiringMap as dependencies necessary for implementing
these changes.